add abstract View class for common UI view stuff
