+++ core/src/main/kotlin/com/github/quillraven/quillycrawler/ui/view/GameView.kt
@@ -1,22 +1,17 @@
 package com.github.quillraven.quillycrawler.ui.view
 
-import com.badlogic.gdx.scenes.scene2d.Stage
 import com.badlogic.gdx.scenes.scene2d.actions.Actions.*
 import com.badlogic.gdx.scenes.scene2d.ui.Label
-import com.badlogic.gdx.scenes.scene2d.ui.Table
 import com.badlogic.gdx.utils.StringBuilder
 import com.github.quillraven.quillycrawler.ui.SkinLabelStyle
-import com.github.quillraven.quillycrawler.ui.model.GameListener
+import com.github.quillraven.quillycrawler.ui.model.GameUiListener
 import com.github.quillraven.quillycrawler.ui.model.GameViewModel
 import ktx.actors.alpha
 import ktx.actors.plusAssign
 import ktx.actors.then
-import ktx.scene2d.KTable
-import ktx.scene2d.Scene2DSkin
 import ktx.scene2d.label
 
-class GameView(private val viewModel: GameViewModel) :
-  Table(Scene2DSkin.defaultSkin), KTable, GameListener {
+class GameView(private val viewModel: GameViewModel) : View(false), GameUiListener {
   private val mapLabel: Label
 
   init {
@@ -30,15 +25,14 @@ class GameView(private val viewModel: GameViewModel) :
     // debugAll()
   }
 
-  override fun setStage(stage: Stage?) {
-    if (stage == null) {
-      // active screen changes away from GameScreen
-      viewModel.removeGameListener(this)
-    } else {
-      // GameScreen becomes active screen
-      viewModel.addGameListener(this)
-    }
-    super.setStage(stage)
+  override fun onShow() {
+    // GameScreen becomes active screen
+    viewModel.addGameListener(this)
+  }
+
+  override fun onHide() {
+    // active screen changes away from GameScreen
+    viewModel.removeGameListener(this)
   }
 
   override fun onMapChange(mapName: StringBuilder) {
